requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.29.0
bcrypt==3.2.2
cffi==1.17.1
click==8.2.1
colorama==0.4.6
fastapi==0.115.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.3.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.32
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.30.6
